+++ package.json
@@ -6,12 +6,12 @@
     "@testing-library/jest-dom": "^5.11.9",
     "@testing-library/react": "^11.2.5",
     "@testing-library/user-event": "^12.8.3",
+    "classnames": "^2.2.6",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-scripts": "4.0.3",
     "web-vitals": "^1.1.0"
   },
-
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
@@ -35,6 +35,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-
+  },
+  "devDependencies": {
+    "react-router-dom": "^5.2.0"
   }
 }
