+++ src/components/MenuHeader/index.js
@@ -1,0 +1,22 @@
+import {useState} from 'react';
+import Menu from "../Menu";
+import Navbar from "../NavBar";
+
+const MenuHeader = () => {
+    const [isMenuActive, setMenuActive] = useState(false);
+    const handleMenuButtonClick = () => {
+        setMenuActive(prevState => !prevState);
+    };
+
+    return (
+        <>
+            <Menu isActive={isMenuActive}/>
+            <Navbar
+                isMenuActive={isMenuActive}
+                changeMenuActive={handleMenuButtonClick}
+            />
+        </>
+    );
+};
+
+export default MenuHeader;
